Switch recommendation API deploy to Modal-only full pipeline.
Remove Render and Hugging Face deploy paths, add modal_app.py, and tighten memory use so production can run bge-base plus the cross-encoder on Modal.

+++ apps/api/scripts/build_index.py
@@ -36,7 +36,7 @@ def main():
 
     print("Saving artifacts...")
     anime_df.to_parquet(artifacts_dir / "anime_df.parquet")
-    np.save(artifacts_dir / "final_features.npy", final_features)
+    np.save(artifacts_dir / "final_features.npy", final_features.astype(np.float32))
     joblib.dump(nn_model, artifacts_dir / "nn_model.pkl")
     joblib.dump(vectorizers, artifacts_dir / "vectorizers.pkl")
